chore(watchAtMost): rename
