+++ packages/shared/index.ts
@@ -39,6 +39,6 @@ export * from './useTimeout'
 export * from './useTimeoutFn'
 export * from './useToggle'
 export * from './utils'
-export * from './watchWithCount'
+export * from './watchAtMost'
 export * from './watchWithFilter'
 export * from './whenever'
